ProtoType for knife Thrower added along with knife prefab
Knife thrower throws a projectile (in this current building only the
knife prefab) at the player location.
Knife is created and flys to the location the player was when it was
spawned.
If it hits the player while traveling, it does x damage (10 currently).
If it doesn't hit player, it lands at the position, waits 2 seconds then
disappears
Knife thrower currently throws a knife every 2 seconds I believe but its
a public variable that can be altered