invincibility periods. Player death counter
Added an invincibility period to both the player and knife thrower once
they have been hit.
when hit, they flash red and are invinicible for 'x' (in this case 1)
second.
added a death counter so that when the player's life reaches 0, the
player is "dying" allowing for the animator to know to using the player
"dying" animation. Once the counter reachs the time (one second for now)
the player is "dead" allowing the animator to know they are dead