Base Enemy and Lancer
Created a Base Enemy Script that all other Enemies will use. (EX:
KnifeThrowerController : BaseEnemy)

Started Work on Lancer Enemy.
Currently if the lancer is attacking (which is set to always be true by
default), he will "charge up" (currently for 1 second) and then move in
the direction that the player was in when he started charging. On
collision with the Player, he deals damage and waits (currently 2
seconds) until he repeats. If he hits an object with the tag "Wall" he
does the same.
I have currently Tested the lancer colliding with the player, and it
seems be be working okay, next I need to test it with walls around.