Small update: Triggers and touch vs attack damage
Added two new variables to the base enemy script. AttackDamage and
TouchDamage

TouchDamage is for when the player walks into an enemy while the enemy
is not attack.
AttackDamage is obviously for when the enemy is attack

Made the Lancer and Knife thrower triggers instead of colliders, so that
the room wont go crazy when a bunch of enemies are in the room colliding
into one another. Also made it so that when the lancer charges the
player, he continues until he hits the wall, meaning he passes through
the player damaging them