Think I have the bear trap working for the enemies
Added a stunned variable and functions to set and change it to the base enemy script, then altered the three enemy scripts to them.

if the lancer hits a trap, he should act like when he hits a wall an become stunned.
if the knifethrower hits a trap, he can keep attacking, unless he is trying to teleport, then he cant attack or teleport.
if the scythe hits the trap, he cant move but should be able to still swing his scythe